debian/copyright fixes
The COPYING file contains the GPL-2, was added by the original author
(Kent Overstreet) and before any other contributions by any other
authors, based on the git history.

Files in the archive contain the single terse statement "GPLv2" and so
together with the COPYING file it can reasonably be implied that the
author intended to release these files under the GNU GPL, version 2.
However, there is no indication that the author intended to release this
under any later version at the option of the recipient, so I have
removed this claim from debian/copyright.

bcache.c is licensed under the PostgreSQL license according to its
header (which is compatible).